Cooperate with cProfile's profile function.
cProfile, apparently, uses the lower-level PyEval_SetProfile instead of
sys.setprofile, so the cooperative installation takes a bit more
work. As a result, theseus tracing is actually a little faster now
because it doesn't go through some extra layers in the general case.

The pure-python Tracer can't cope with this, but can at least provide an
informative error message.